requirements.txt
flask==3.1.0
flask-cors==5.0.0
python-dotenv==1.0.1
groq==0.13.0
gunicorn==23.0.0
flask-sqlalchemy==3.1.0
